+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 ## [Unreleased][]
 
+### Fixed
+
+-   Handle non-primitive replacers
+
 ## [1.0.0][] - 2020-12-15
 
 ### Added
 
 -   Initial implementation
 
-[unreleased]:
-	https://github.com/niksy/string-replace-all-ponyfill/compare/v1.0.0...HEAD
 [1.0.0]: https://github.com/niksy/string-replace-all-ponyfill/tree/v1.0.0