+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased][]
 
+## [1.0.1][] - 2020-12-17
+
 ### Added
 
 -   Named export `preferNative`
@@ -18,3 +20,6 @@
 -   Initial implementation
 
 [1.0.0]: https://github.com/niksy/string-replace-all-ponyfill/tree/v1.0.0
+[unreleased]:
+	https://github.com/niksy/string-replace-all-ponyfill/compare/v1.0.1...HEAD
+[1.0.1]: https://github.com/niksy/string-replace-all-ponyfill/tree/v1.0.1